Add prospect owner/watch features and split admin routes
Features:
- Show "Me!" for prospects current user owns, otherwise show owner name
- Add owner assignment dropdown to prospects page
- Add watch/unwatch toggle (star icon) for prospects
- Display linked domains, Slack user counts, last activity
- Show open invoices with badge linking to Stripe
- Add "Open Invoices" view tab

Refactoring:
- Split admin.ts (3830 lines) into focused modules:
  - admin/prospects.ts: Prospect listing, creation, updates
  - admin/organizations.ts: Org details, activities, stakeholders
  - admin/enrichment.ts: Lusha enrichment, company search
  - admin/domains.ts: Domain discovery, email contacts
  - admin/cleanup.ts: AI-powered prospect cleanup
- Main admin.ts now ~520 lines composing route modules
