+++ server/src/routes/admin.ts
@@ -1,46 +1,29 @@
 /**
  * Admin routes module
  *
- * This module contains admin-only routes extracted from http.ts.
- * New admin routes should be added here to keep http.ts from growing.
- *
- * Existing admin routes in http.ts will be migrated here incrementally.
+ * This module composes admin routes from individual route modules.
+ * Routes are organized into focused modules for better maintainability.
  */
 
 import { Router } from "express";
-import { WorkOS, DomainDataState } from "@workos-inc/node";
+import { WorkOS } from "@workos-inc/node";
 import { getPool } from "../db/client.js";
 import { createLogger } from "../logger.js";
 import { requireAuth, requireAdmin } from "../middleware/auth.js";
 import { serveHtmlWithConfig } from "../utils/html-config.js";
-import { SlackDatabase } from "../db/slack-db.js";
-import { OrganizationDatabase } from "../db/organization-db.js";
+import { getMemberContext, getWebMemberContext } from "../addie/member-context.js";
 import {
   createCheckoutSession,
   getProductsForCustomer,
-  getPendingInvoices,
   createAndSendInvoice,
 } from "../billing/stripe-client.js";
-import { getMemberContext, getWebMemberContext } from "../addie/member-context.js";
-import {
-  getLushaClient,
-  isLushaConfigured,
-  mapIndustryToCompanyType,
-  formatRevenueRange,
-} from "../services/lusha.js";
-import {
-  enrichOrganization,
-  enrichDomainsInBatch,
-  autoEnrichOrganization,
-} from "../services/enrichment.js";
-import { createProspect } from "../services/prospect.js";
-import {
-  getCleanupService,
-  isCleanupConfigured,
-} from "../services/prospect-cleanup.js";
 
-const slackDb = new SlackDatabase();
-const orgDb = new OrganizationDatabase();
+// Import route modules
+import { setupProspectRoutes } from "./admin/prospects.js";
+import { setupOrganizationRoutes } from "./admin/organizations.js";
+import { setupEnrichmentRoutes } from "./admin/enrichment.js";
+import { setupDomainRoutes } from "./admin/domains.js";
+import { setupCleanupRoutes } from "./admin/cleanup.js";
 
 const logger = createLogger("admin-routes");
 
@@ -85,2854 +68,29 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
   });
 
   // =========================================================================
-  // PROSPECT MANAGEMENT API (mounted at /api/admin)
+  // SET UP ROUTE MODULES
   // =========================================================================
 
-  // GET /api/admin/prospects - List all prospects with action-based views
-  apiRouter.get("/prospects", requireAuth, requireAdmin, async (req, res) => {
-    try {
-      const pool = getPool();
-      const { status, source, view, owner } = req.query;
+  // Prospect management routes
+  setupProspectRoutes(apiRouter, { workos });
 
-      // Base SELECT fields
-      const selectFields = `
-        SELECT
-          o.workos_organization_id,
-          o.name,
-          o.company_type,
-          o.revenue_tier,
-          o.is_personal,
-          COALESCE(o.prospect_status, 'signed_up') as prospect_status,
-          COALESCE(o.prospect_source, 'organic') as prospect_source,
-          o.prospect_owner,
-          o.prospect_notes,
-          o.prospect_contact_name,
-          o.prospect_contact_email,
-          o.prospect_contact_title,
-          o.prospect_next_action,
-          o.prospect_next_action_date,
-          o.parent_organization_id,
-          o.created_at,
-          o.updated_at,
-          o.invoice_requested_at,
-          o.last_activity_at,
-          p.name as parent_name,
-          (SELECT COUNT(*) FROM organizations WHERE parent_organization_id = o.workos_organization_id) as subsidiary_count,
-          o.subscription_status,
-          o.subscription_product_name,
-          o.subscription_current_period_end
-      `;
+  // Organization detail and management routes
+  setupOrganizationRoutes(pageRouter, apiRouter, { workos });
 
-      const params: (string | Date | null)[] = [];
-      let query = "";
-      let orderBy = "";
+  // Company enrichment and prospecting routes
+  setupEnrichmentRoutes(apiRouter);
 
-      // Action-based views
-      if (view && typeof view === "string") {
-        switch (view) {
-          case "needs_followup":
-            // Orgs with pending next steps due in next 7 days
-            query = `
-              ${selectFields},
-              na.next_step_due_date as followup_due,
-              na.description as followup_description
-              FROM organizations o
-              LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-              INNER JOIN org_activities na ON na.organization_id = o.workos_organization_id
-                AND na.is_next_step = TRUE
-                AND na.next_step_completed_at IS NULL
-                AND (na.next_step_due_date IS NULL OR na.next_step_due_date <= NOW() + INTERVAL '7 days')
-            `;
-            orderBy = ` ORDER BY na.next_step_due_date ASC NULLS FIRST`;
-            break;
+  // Domain discovery, email contacts, and org domains routes
+  setupDomainRoutes(apiRouter, { workos });
 
-          case "hot_prospects":
-            // Non-paying orgs with high engagement (level 3+)
-            // We'll calculate engagement in JS, so just get non-paying orgs
-            query = `
-              ${selectFields}
-              FROM organizations o
-              LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-              WHERE (
-                o.subscription_status IS NULL
-                OR o.subscription_status NOT IN ('active', 'trialing')
-                OR o.subscription_canceled_at IS NOT NULL
-              )
-            `;
-            orderBy = ` ORDER BY o.invoice_requested_at DESC NULLS LAST, o.last_activity_at DESC NULLS LAST`;
-            break;
-
-          case "new_signups":
-            // Orgs created in last 14 days with no activities logged
-            const fourteenDaysAgo = new Date();
-            fourteenDaysAgo.setDate(fourteenDaysAgo.getDate() - 14);
-            params.push(fourteenDaysAgo);
-            query = `
-              ${selectFields}
-              FROM organizations o
-              LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-              WHERE o.created_at > $${params.length}
-                AND NOT EXISTS (SELECT 1 FROM org_activities WHERE organization_id = o.workos_organization_id)
-            `;
-            orderBy = ` ORDER BY o.created_at DESC`;
-            break;
-
-          case "going_cold":
-            // Orgs with activity more than 30 days ago (but had some activity)
-            const thirtyDaysAgo = new Date();
-            thirtyDaysAgo.setDate(thirtyDaysAgo.getDate() - 30);
-            params.push(thirtyDaysAgo);
-            query = `
-              ${selectFields}
-              FROM organizations o
-              LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-              WHERE o.last_activity_at IS NOT NULL
-                AND o.last_activity_at < $${params.length}
-                AND (
-                  o.subscription_status IS NULL
-                  OR o.subscription_status NOT IN ('active', 'trialing')
-                  OR o.subscription_canceled_at IS NOT NULL
-                )
-            `;
-            orderBy = ` ORDER BY o.last_activity_at ASC`;
-            break;
-
-          case "renewals":
-            // Paying orgs with subscription ending in next 60 days
-            const sixtyDaysFromNow = new Date();
-            sixtyDaysFromNow.setDate(sixtyDaysFromNow.getDate() + 60);
-            params.push(new Date());
-            params.push(sixtyDaysFromNow);
-            query = `
-              ${selectFields}
-              FROM organizations o
-              LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-              WHERE o.subscription_status = 'active'
-                AND o.subscription_current_period_end IS NOT NULL
-                AND o.subscription_current_period_end >= $${params.length - 1}
-                AND o.subscription_current_period_end <= $${params.length}
-            `;
-            orderBy = ` ORDER BY o.subscription_current_period_end ASC`;
-            break;
-
-          case "low_engagement":
-            // Paying members with low engagement (we'll filter by engagement_level <= 2 in JS)
-            query = `
-              ${selectFields}
-              FROM organizations o
-              LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-              WHERE o.subscription_status = 'active'
-                AND o.subscription_canceled_at IS NULL
-            `;
-            orderBy = ` ORDER BY o.last_activity_at ASC NULLS FIRST`;
-            break;
-
-          case "my_accounts":
-            // Accounts where current user is a stakeholder
-            const userId = req.user?.id;
-            if (userId) {
-              params.push(userId);
-              query = `
-                ${selectFields},
-                os.role as stakeholder_role
-                FROM organizations o
-                LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-                INNER JOIN org_stakeholders os ON os.organization_id = o.workos_organization_id
-                  AND os.user_id = $${params.length}
-              `;
-              orderBy = ` ORDER BY os.role ASC, o.name ASC`;
-            } else {
-              // No user ID, return empty
-              return res.json([]);
-            }
-            break;
-
-          default:
-            // Fall back to default query for unknown views
-            query = `
-              ${selectFields}
-              FROM organizations o
-              LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-              WHERE (
-                o.subscription_status IS NULL
-                OR o.subscription_status NOT IN ('active', 'trialing')
-                OR o.subscription_canceled_at IS NOT NULL
-              )
-            `;
-            orderBy = ` ORDER BY o.created_at DESC`;
-        }
-      } else {
-        // Default: Show all non-paying orgs (original behavior)
-        query = `
-          ${selectFields}
-          FROM organizations o
-          LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-          WHERE (
-            o.subscription_status IS NULL
-            OR o.subscription_status NOT IN ('active', 'trialing')
-            OR o.subscription_canceled_at IS NOT NULL
-          )
-        `;
-        orderBy = ` ORDER BY o.prospect_next_action_date ASC NULLS LAST, o.created_at DESC`;
-      }
-
-      // Apply additional filters (status, source, owner) if not using a specialized view
-      if (!view || view === "hot_prospects" || view === "going_cold" || view === "low_engagement") {
-        if (status && typeof status === "string") {
-          params.push(status);
-          query += ` AND o.prospect_status = $${params.length}`;
-        }
-
-        if (source && typeof source === "string") {
-          params.push(source);
-          query += ` AND o.prospect_source = $${params.length}`;
-        }
-
-        if (owner && typeof owner === "string") {
-          params.push(owner);
-          query += ` AND o.prospect_owner = $${params.length}`;
-        }
-      }
-
-      query += orderBy;
-
-      const result = await pool.query(query, params);
-
-      // Get working group counts for all orgs
-      const workingGroupCounts = await pool.query(`
-        SELECT workos_organization_id as organization_id, COUNT(DISTINCT working_group_id) as wg_count
-        FROM working_group_memberships
-        WHERE status = 'active' AND workos_organization_id IS NOT NULL
-        GROUP BY workos_organization_id
-      `);
-      const wgCountMap = new Map(
-        workingGroupCounts.rows.map((r) => [r.organization_id, parseInt(r.wg_count)])
-      );
-
-      // Get recent activity counts (last 30 days)
-      const thirtyDaysAgo = new Date();
-      thirtyDaysAgo.setDate(thirtyDaysAgo.getDate() - 30);
-      const recentActivityCounts = await pool.query(
-        `
-        SELECT organization_id, COUNT(*) as activity_count
-        FROM org_activities
-        WHERE activity_date > $1
-        GROUP BY organization_id
-      `,
-        [thirtyDaysAgo]
-      );
-      const activityCountMap = new Map(
-        recentActivityCounts.rows.map((r) => [
-          r.organization_id,
-          parseInt(r.activity_count),
-        ])
-      );
-
-      // Enrich with WorkOS membership count and engagement level
-      const prospects = await Promise.all(
-        result.rows.map(async (row) => {
-          let memberCount = 0;
-          try {
-            if (workos) {
-              const memberships =
-                await workos.userManagement.listOrganizationMemberships({
-                  organizationId: row.workos_organization_id,
-                });
-              memberCount = memberships.data?.length || 0;
-            }
-          } catch {
-            // Org might not exist in WorkOS yet or other error
-          }
-
-          // Calculate engagement level
-          const wgCount = wgCountMap.get(row.workos_organization_id) || 0;
-          const recentActivityCount =
-            activityCountMap.get(row.workos_organization_id) || 0;
-
-          let engagementLevel = 1; // Base level - exists
-          const engagementReasons: string[] = [];
-
-          if (row.invoice_requested_at) {
-            engagementLevel = 5;
-            engagementReasons.push("Requested invoice");
-          } else if (wgCount > 0) {
-            engagementLevel = 4;
-            engagementReasons.push(`In ${wgCount} working group(s)`);
-          } else if (memberCount > 0) {
-            engagementLevel = 3;
-            engagementReasons.push(`${memberCount} team member(s)`);
-          } else if (recentActivityCount > 0) {
-            engagementLevel = 2;
-            engagementReasons.push("Recent contact");
-          }
-
-          return {
-            ...row,
-            member_count: memberCount,
-            has_members: memberCount > 0,
-            working_group_count: wgCount,
-            engagement_level: engagementLevel,
-            engagement_reasons: engagementReasons,
-          };
-        })
-      );
-
-      // Filter by engagement level for specific views
-      let filteredProspects = prospects;
-      if (view === "hot_prospects") {
-        // Only show high engagement (level 3+)
-        filteredProspects = prospects.filter((p) => p.engagement_level >= 3);
-      } else if (view === "low_engagement") {
-        // Only show low engagement (level 2 or less)
-        filteredProspects = prospects.filter((p) => p.engagement_level <= 2);
-      }
-
-      res.json(filteredProspects);
-    } catch (error) {
-      logger.error({ err: error }, "Error fetching prospects");
-      res.status(500).json({
-        error: "Internal server error",
-        message: "Unable to fetch prospects",
-      });
-    }
-  });
-
-  // POST /api/admin/prospects - Create a new prospect
-  apiRouter.post("/prospects", requireAuth, requireAdmin, async (req, res) => {
-    try {
-      const {
-        name,
-        domain,
-        company_type,
-        prospect_status,
-        prospect_source,
-        prospect_notes,
-        prospect_contact_name,
-        prospect_contact_email,
-        prospect_contact_title,
-        prospect_next_action,
-        prospect_next_action_date,
-        prospect_owner,
-        parent_organization_id,
-      } = req.body;
-
-      if (!name || typeof name !== "string") {
-        return res.status(400).json({ error: "Company name is required" });
-      }
-
-      // Use centralized prospect service
-      const result = await createProspect({
-        name,
-        domain,
-        company_type,
-        prospect_status,
-        prospect_source: prospect_source || "aao_launch_list",
-        prospect_notes,
-        prospect_contact_name,
-        prospect_contact_email,
-        prospect_contact_title,
-        prospect_next_action,
-        prospect_next_action_date,
-        prospect_owner,
-        parent_organization_id,
-      });
-
-      if (!result.success) {
-        if (result.alreadyExists) {
-          return res.status(409).json({
-            error: "Organization already exists",
-            message: result.error,
-            organization: result.organization,
-          });
-        }
-        return res.status(400).json({
-          error: "Failed to create prospect",
-          message: result.error,
-        });
-      }
-
-      res.status(201).json(result.organization);
-    } catch (error) {
-      logger.error({ err: error }, "Error creating prospect");
-      res.status(500).json({
-        error: "Internal server error",
-        message: "Unable to create prospect",
-      });
-    }
-  });
-
-  // PUT /api/admin/prospects/:orgId - Update prospect
-  apiRouter.put(
-    "/prospects/:orgId",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const {
-          prospect_status,
-          prospect_source,
-          prospect_owner,
-          prospect_notes,
-          prospect_contact_name,
-          prospect_contact_email,
-          prospect_contact_title,
-          prospect_next_action,
-          prospect_next_action_date,
-          company_type,
-          revenue_tier,
-          parent_organization_id,
-        } = req.body;
-
-        const pool = getPool();
-
-        // Build dynamic update query
-        const updates: string[] = [];
-        const values: any[] = [];
-        let paramIndex = 1;
-
-        if (prospect_status !== undefined) {
-          updates.push(`prospect_status = $${paramIndex++}`);
-          values.push(prospect_status);
-        }
-        if (prospect_source !== undefined) {
-          updates.push(`prospect_source = $${paramIndex++}`);
-          values.push(prospect_source);
-        }
-        if (prospect_owner !== undefined) {
-          updates.push(`prospect_owner = $${paramIndex++}`);
-          values.push(prospect_owner);
-        }
-        if (prospect_notes !== undefined) {
-          updates.push(`prospect_notes = $${paramIndex++}`);
-          values.push(prospect_notes);
-        }
-        if (prospect_contact_name !== undefined) {
-          updates.push(`prospect_contact_name = $${paramIndex++}`);
-          values.push(prospect_contact_name);
-        }
-        if (prospect_contact_email !== undefined) {
-          updates.push(`prospect_contact_email = $${paramIndex++}`);
-          values.push(prospect_contact_email);
-        }
-        if (prospect_contact_title !== undefined) {
-          updates.push(`prospect_contact_title = $${paramIndex++}`);
-          values.push(prospect_contact_title);
-        }
-        if (prospect_next_action !== undefined) {
-          updates.push(`prospect_next_action = $${paramIndex++}`);
-          values.push(prospect_next_action);
-        }
-        if (prospect_next_action_date !== undefined) {
-          updates.push(`prospect_next_action_date = $${paramIndex++}`);
-          values.push(prospect_next_action_date);
-        }
-        if (company_type !== undefined) {
-          updates.push(`company_type = $${paramIndex++}`);
-          values.push(company_type);
-        }
-        if (revenue_tier !== undefined) {
-          updates.push(`revenue_tier = $${paramIndex++}`);
-          values.push(revenue_tier);
-        }
-        if (parent_organization_id !== undefined) {
-          updates.push(`parent_organization_id = $${paramIndex++}`);
-          values.push(parent_organization_id);
-        }
-
-        if (updates.length === 0) {
-          return res.status(400).json({ error: "No fields to update" });
-        }
-
-        updates.push("updated_at = NOW()");
-        values.push(orgId);
-
-        const result = await pool.query(
-          `UPDATE organizations
-         SET ${updates.join(", ")}
-         WHERE workos_organization_id = $${paramIndex}
-         RETURNING *`,
-          values
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({ error: "Prospect not found" });
-        }
-
-        res.json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, "Error updating prospect");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to update prospect",
-        });
-      }
-    }
-  );
-
-  // GET /api/admin/prospects/stats - Get prospect statistics
-  apiRouter.get(
-    "/prospects/stats",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const pool = getPool();
-
-        // Count all non-paying orgs by status (including signed_up for those without explicit status)
-        const result = await pool.query(`
-        SELECT
-          COALESCE(prospect_status, 'signed_up') as prospect_status,
-          COUNT(*) as count
-        FROM organizations
-        WHERE (
-          subscription_status IS NULL
-          OR subscription_status NOT IN ('active', 'trialing')
-          OR subscription_canceled_at IS NOT NULL
-        )
-        GROUP BY COALESCE(prospect_status, 'signed_up')
-        ORDER BY
-          CASE COALESCE(prospect_status, 'signed_up')
-            WHEN 'signed_up' THEN 0
-            WHEN 'prospect' THEN 1
-            WHEN 'contacted' THEN 2
-            WHEN 'interested' THEN 3
-            WHEN 'negotiating' THEN 4
-            WHEN 'converted' THEN 5
-            WHEN 'declined' THEN 6
-            ELSE 7
-          END
-      `);
-
-        const stats: Record<string, number> = {};
-        let total = 0;
-
-        for (const row of result.rows) {
-          stats[row.prospect_status] = parseInt(row.count);
-          total += parseInt(row.count);
-        }
-
-        res.json({
-          by_status: stats,
-          total,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching prospect stats");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch prospect statistics",
-        });
-      }
-    }
-  );
-
-  // GET /api/admin/organizations - List all organizations (for parent org dropdown)
-  apiRouter.get(
-    "/organizations",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const pool = getPool();
-
-        const result = await pool.query(`
-        SELECT
-          workos_organization_id,
-          name,
-          company_type,
-          prospect_status
-        FROM organizations
-        ORDER BY name ASC
-      `);
-
-        res.json(result.rows);
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching organizations");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch organizations",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // ORGANIZATION DETAIL PAGE AND API
-  // =========================================================================
-
-  // Page route for org detail
-  pageRouter.get(
-    "/organizations/:orgId",
-    requireAuth,
-    requireAdmin,
-    (req, res) => {
-      serveHtmlWithConfig(req, res, "admin-org-detail.html").catch((err) => {
-        logger.error({ err }, "Error serving admin org detail page");
-        res.status(500).send("Internal server error");
-      });
-    }
-  );
-
-  // GET /api/admin/organizations/:orgId - Get full org details with engagement data
-  apiRouter.get(
-    "/organizations/:orgId",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const pool = getPool();
-
-        // Get organization details
-        const orgResult = await pool.query(
-          `
-          SELECT
-            o.*,
-            p.name as parent_name,
-            (SELECT COUNT(*) FROM organizations WHERE parent_organization_id = o.workos_organization_id) as subsidiary_count
-          FROM organizations o
-          LEFT JOIN organizations p ON o.parent_organization_id = p.workos_organization_id
-          WHERE o.workos_organization_id = $1
-        `,
-          [orgId]
-        );
-
-        if (orgResult.rows.length === 0) {
-          return res.status(404).json({ error: "Organization not found" });
-        }
-
-        const org = orgResult.rows[0];
-
-        // Get member count from WorkOS
-        let memberCount = 0;
-        let members: any[] = [];
-        try {
-          if (workos) {
-            const memberships =
-              await workos.userManagement.listOrganizationMemberships({
-                organizationId: orgId,
-              });
-            memberCount = memberships.data?.length || 0;
-
-            // Get user details for each membership
-            for (const membership of memberships.data || []) {
-              try {
-                const user = await workos.userManagement.getUser(
-                  membership.userId
-                );
-                members.push({
-                  id: user.id,
-                  email: user.email,
-                  firstName: user.firstName,
-                  lastName: user.lastName,
-                  role: membership.role?.slug || "member",
-                });
-              } catch {
-                // User might not exist
-              }
-            }
-          }
-        } catch {
-          // Org might not exist in WorkOS
-        }
-
-        // Get working group memberships
-        const workingGroupResult = await pool.query(
-          `
-          SELECT DISTINCT wg.id, wg.name, wg.slug, wgm.status, wgm.joined_at
-          FROM working_group_memberships wgm
-          JOIN working_groups wg ON wgm.working_group_id = wg.id
-          WHERE wgm.workos_organization_id = $1 AND wgm.status = 'active'
-        `,
-          [orgId]
-        );
-
-        // Get recent activities
-        const activitiesResult = await pool.query(
-          `
-          SELECT *
-          FROM org_activities
-          WHERE organization_id = $1
-          ORDER BY activity_date DESC
-          LIMIT 50
-        `,
-          [orgId]
-        );
-
-        // Get pending next steps
-        const nextStepsResult = await pool.query(
-          `
-          SELECT *
-          FROM org_activities
-          WHERE organization_id = $1
-            AND is_next_step = TRUE
-            AND next_step_completed_at IS NULL
-          ORDER BY next_step_due_date ASC NULLS LAST
-        `,
-          [orgId]
-        );
-
-        // Get engagement signals using the new engagement tracking system
-        const engagementSignals = await orgDb.getEngagementSignals(orgId);
-
-        // Calculate engagement level based on signals
-        let engagementLevel = 1; // Base level - exists
-        let engagementReasons: string[] = [];
-
-        // Priority-based scoring - use human interest level first if set
-        if (engagementSignals.interest_level === 'very_high') {
-          engagementLevel = 5;
-          engagementReasons.push(`Interest: Very High (${engagementSignals.interest_level_set_by || 'admin'})`);
-        } else if (engagementSignals.interest_level === 'high') {
-          engagementLevel = 4;
-          engagementReasons.push(`Interest: High (${engagementSignals.interest_level_set_by || 'admin'})`);
-        } else if (org.invoice_requested_at) {
-          engagementLevel = 5;
-          engagementReasons.push("Requested invoice");
-        } else if (engagementSignals.working_group_count > 0) {
-          engagementLevel = 4;
-          engagementReasons.push(`In ${engagementSignals.working_group_count} working group(s)`);
-        } else if (engagementSignals.has_member_profile) {
-          engagementLevel = 4;
-          engagementReasons.push("Member profile configured");
-        } else if (engagementSignals.login_count_30d > 3) {
-          engagementLevel = 3;
-          engagementReasons.push(`${engagementSignals.login_count_30d} dashboard logins (30d)`);
-        } else if (memberCount > 0) {
-          engagementLevel = 3;
-          engagementReasons.push(`${memberCount} team member(s)`);
-        } else if (engagementSignals.email_click_count_30d > 0) {
-          engagementLevel = 2;
-          engagementReasons.push(`${engagementSignals.email_click_count_30d} email clicks (30d)`);
-        } else if (engagementSignals.login_count_30d > 0) {
-          engagementLevel = 2;
-          engagementReasons.push(`${engagementSignals.login_count_30d} dashboard login(s) (30d)`);
-        } else if (activitiesResult.rows.length > 0) {
-          const recentActivity = activitiesResult.rows.find((a) => {
-            const activityDate = new Date(a.activity_date);
-            const thirtyDaysAgo = new Date();
-            thirtyDaysAgo.setDate(thirtyDaysAgo.getDate() - 30);
-            return activityDate > thirtyDaysAgo;
-          });
-          if (recentActivity) {
-            engagementLevel = 2;
-            engagementReasons.push("Recent contact");
-          }
-        }
-
-        // Handle low/medium interest levels - should cap the engagement
-        if (engagementSignals.interest_level === 'low') {
-          engagementLevel = Math.min(engagementLevel, 2);
-          engagementReasons.unshift(`Interest: Low (${engagementSignals.interest_level_set_by || 'admin'})`);
-        } else if (engagementSignals.interest_level === 'medium') {
-          engagementLevel = Math.min(engagementLevel, 3);
-          engagementReasons.unshift(`Interest: Medium (${engagementSignals.interest_level_set_by || 'admin'})`);
-        }
-
-        // Fetch pending invoices if org has a Stripe customer ID
-        let pendingInvoices: Awaited<ReturnType<typeof getPendingInvoices>> = [];
-        if (org.stripe_customer_id) {
-          try {
-            pendingInvoices = await getPendingInvoices(org.stripe_customer_id);
-          } catch (err) {
-            logger.warn({ err, orgId, stripeCustomerId: org.stripe_customer_id }, 'Error fetching pending invoices');
-          }
-        }
-
-        res.json({
-          ...org,
-          member_count: memberCount,
-          members,
-          working_groups: workingGroupResult.rows,
-          activities: activitiesResult.rows,
-          next_steps: nextStepsResult.rows,
-          engagement_level: engagementLevel,
-          engagement_reasons: engagementReasons,
-          engagement_signals: engagementSignals,
-          pending_invoices: pendingInvoices,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching organization details");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch organization details",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/organizations/:orgId/activities - Log an activity
-  apiRouter.post(
-    "/organizations/:orgId/activities",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const {
-          activity_type,
-          description,
-          activity_date,
-          is_next_step,
-          next_step_due_date,
-          next_step_owner_user_id,
-          next_step_owner_name,
-        } = req.body;
-
-        if (!activity_type) {
-          return res.status(400).json({ error: "activity_type is required" });
-        }
-
-        const pool = getPool();
-
-        // Get logged-in user info
-        const loggedByUserId = req.user?.id || null;
-        const loggedByName = req.user
-          ? `${req.user.firstName || ""} ${req.user.lastName || ""}`.trim() ||
-            req.user.email
-          : null;
-
-        const result = await pool.query(
-          `
-          INSERT INTO org_activities (
-            organization_id,
-            activity_type,
-            description,
-            logged_by_user_id,
-            logged_by_name,
-            activity_date,
-            is_next_step,
-            next_step_due_date,
-            next_step_owner_user_id,
-            next_step_owner_name
-          ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
-          RETURNING *
-        `,
-          [
-            orgId,
-            activity_type,
-            description || null,
-            loggedByUserId,
-            loggedByName,
-            activity_date || new Date(),
-            is_next_step || false,
-            next_step_due_date || null,
-            next_step_owner_user_id || null,
-            next_step_owner_name || null,
-          ]
-        );
-
-        // Update last_activity_at on the organization
-        await pool.query(
-          `
-          UPDATE organizations
-          SET last_activity_at = $2, updated_at = NOW()
-          WHERE workos_organization_id = $1
-        `,
-          [orgId, activity_date || new Date()]
-        );
-
-        // If invoice_requested, update that field too
-        if (activity_type === "invoice_requested") {
-          await pool.query(
-            `
-            UPDATE organizations
-            SET invoice_requested_at = $2
-            WHERE workos_organization_id = $1 AND invoice_requested_at IS NULL
-          `,
-            [orgId, activity_date || new Date()]
-          );
-        }
-
-        res.status(201).json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, "Error logging activity");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to log activity",
-        });
-      }
-    }
-  );
-
-  // PUT /api/admin/organizations/:orgId/activities/:activityId - Update activity (e.g., complete next step)
-  apiRouter.put(
-    "/organizations/:orgId/activities/:activityId",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId, activityId } = req.params;
-        const { next_step_completed_at, description, next_step_due_date } =
-          req.body;
-
-        const pool = getPool();
-
-        const updates: string[] = [];
-        const values: any[] = [];
-        let paramIndex = 1;
-
-        if (next_step_completed_at !== undefined) {
-          updates.push(`next_step_completed_at = $${paramIndex++}`);
-          values.push(next_step_completed_at);
-        }
-        if (description !== undefined) {
-          updates.push(`description = $${paramIndex++}`);
-          values.push(description);
-        }
-        if (next_step_due_date !== undefined) {
-          updates.push(`next_step_due_date = $${paramIndex++}`);
-          values.push(next_step_due_date);
-        }
-
-        if (updates.length === 0) {
-          return res.status(400).json({ error: "No fields to update" });
-        }
-
-        updates.push("updated_at = NOW()");
-        values.push(activityId);
-        values.push(orgId);
-
-        const result = await pool.query(
-          `
-          UPDATE org_activities
-          SET ${updates.join(", ")}
-          WHERE id = $${paramIndex} AND organization_id = $${paramIndex + 1}
-          RETURNING *
-        `,
-          values
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({ error: "Activity not found" });
-        }
-
-        res.json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, "Error updating activity");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to update activity",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // STAKEHOLDER MANAGEMENT API
-  // =========================================================================
-
-  // GET /api/admin/organizations/:orgId/stakeholders - Get all stakeholders for an org
-  apiRouter.get(
-    "/organizations/:orgId/stakeholders",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const pool = getPool();
-
-        const result = await pool.query(
-          `
-          SELECT *
-          FROM org_stakeholders
-          WHERE organization_id = $1
-          ORDER BY
-            CASE role
-              WHEN 'owner' THEN 1
-              WHEN 'interested' THEN 2
-              WHEN 'connected' THEN 3
-            END,
-            created_at ASC
-        `,
-          [orgId]
-        );
-
-        res.json(result.rows);
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching stakeholders");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch stakeholders",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/organizations/:orgId/stakeholders - Add stakeholder (or update role)
-  apiRouter.post(
-    "/organizations/:orgId/stakeholders",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const { user_id, user_name, user_email, role, notes } = req.body;
-
-        // If no user_id provided, use the current logged-in user
-        const actualUserId = user_id || req.user?.id;
-        const actualUserName =
-          user_name ||
-          (req.user
-            ? `${req.user.firstName || ""} ${req.user.lastName || ""}`.trim() ||
-              req.user.email
-            : null);
-        const actualUserEmail = user_email || req.user?.email;
-
-        if (!actualUserId) {
-          return res.status(400).json({ error: "user_id is required" });
-        }
-
-        if (!role || !["owner", "interested", "connected"].includes(role)) {
-          return res.status(400).json({
-            error: "role must be one of: owner, interested, connected",
-          });
-        }
-
-        const pool = getPool();
-
-        // Upsert: insert or update if already exists
-        const result = await pool.query(
-          `
-          INSERT INTO org_stakeholders (
-            organization_id,
-            user_id,
-            user_name,
-            user_email,
-            role,
-            notes
-          ) VALUES ($1, $2, $3, $4, $5, $6)
-          ON CONFLICT (organization_id, user_id)
-          DO UPDATE SET
-            role = EXCLUDED.role,
-            notes = COALESCE(EXCLUDED.notes, org_stakeholders.notes),
-            user_name = EXCLUDED.user_name,
-            user_email = EXCLUDED.user_email,
-            updated_at = NOW()
-          RETURNING *
-        `,
-          [
-            orgId,
-            actualUserId,
-            actualUserName,
-            actualUserEmail,
-            role,
-            notes || null,
-          ]
-        );
-
-        res.status(201).json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, "Error adding stakeholder");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to add stakeholder",
-        });
-      }
-    }
-  );
-
-  // DELETE /api/admin/organizations/:orgId/stakeholders/:stakeholderId - Remove stakeholder
-  apiRouter.delete(
-    "/organizations/:orgId/stakeholders/:stakeholderId",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId, stakeholderId } = req.params;
-        const pool = getPool();
-
-        const result = await pool.query(
-          `
-          DELETE FROM org_stakeholders
-          WHERE id = $1 AND organization_id = $2
-          RETURNING *
-        `,
-          [stakeholderId, orgId]
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({ error: "Stakeholder not found" });
-        }
-
-        res.json({ success: true, deleted: result.rows[0] });
-      } catch (error) {
-        logger.error({ err: error }, "Error removing stakeholder");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to remove stakeholder",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/organizations/:orgId/stakeholders/me - Quick "I'm connected" for current user
-  apiRouter.post(
-    "/organizations/:orgId/stakeholders/me",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const { role } = req.body;
-
-        if (!req.user?.id) {
-          return res.status(401).json({ error: "User not authenticated" });
-        }
-
-        const actualRole = role || "connected";
-        if (!["owner", "interested", "connected"].includes(actualRole)) {
-          return res.status(400).json({
-            error: "role must be one of: owner, interested, connected",
-          });
-        }
-
-        const pool = getPool();
-
-        const userName =
-          `${req.user.firstName || ""} ${req.user.lastName || ""}`.trim() ||
-          req.user.email;
-
-        // Upsert for current user
-        const result = await pool.query(
-          `
-          INSERT INTO org_stakeholders (
-            organization_id,
-            user_id,
-            user_name,
-            user_email,
-            role
-          ) VALUES ($1, $2, $3, $4, $5)
-          ON CONFLICT (organization_id, user_id)
-          DO UPDATE SET
-            role = EXCLUDED.role,
-            updated_at = NOW()
-          RETURNING *
-        `,
-          [orgId, req.user.id, userName, req.user.email, actualRole]
-        );
-
-        res.status(201).json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, "Error adding self as stakeholder");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to add yourself as stakeholder",
-        });
-      }
-    }
-  );
-
-  // DELETE /api/admin/organizations/:orgId/stakeholders/me - Remove self as stakeholder
-  apiRouter.delete(
-    "/organizations/:orgId/stakeholders/me",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-
-        if (!req.user?.id) {
-          return res.status(401).json({ error: "User not authenticated" });
-        }
-
-        const pool = getPool();
-
-        const result = await pool.query(
-          `
-          DELETE FROM org_stakeholders
-          WHERE organization_id = $1 AND user_id = $2
-          RETURNING *
-        `,
-          [orgId, req.user.id]
-        );
-
-        if (result.rows.length === 0) {
-          return res
-            .status(404)
-            .json({ error: "You are not a stakeholder for this organization" });
-        }
-
-        res.json({ success: true, deleted: result.rows[0] });
-      } catch (error) {
-        logger.error({ err: error }, "Error removing self as stakeholder");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to remove yourself as stakeholder",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // ENGAGEMENT / INTEREST LEVEL MANAGEMENT
-  // =========================================================================
-
-  // PUT /api/admin/organizations/:orgId/interest-level - Set interest level for an org
-  apiRouter.put(
-    "/organizations/:orgId/interest-level",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const { interest_level, note } = req.body;
-
-        // Validate interest level
-        const validLevels = ['low', 'medium', 'high', 'very_high', null];
-        if (!validLevels.includes(interest_level)) {
-          return res.status(400).json({
-            error: "Invalid interest_level. Must be one of: low, medium, high, very_high (or null to clear)",
-          });
-        }
-
-        // Get the admin's name
-        const setBy = req.user
-          ? `${req.user.firstName || ""} ${req.user.lastName || ""}`.trim() ||
-            req.user.email
-          : "admin";
-
-        await orgDb.setInterestLevel(orgId, {
-          interest_level,
-          note,
-          set_by: setBy,
-        });
-
-        // Return the updated engagement signals
-        const engagementSignals = await orgDb.getEngagementSignals(orgId);
-
-        logger.info(
-          { orgId, interest_level, setBy },
-          "Interest level updated"
-        );
-
-        res.json({
-          success: true,
-          engagement_signals: engagementSignals,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error setting interest level");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to set interest level",
-        });
-      }
-    }
-  );
-
-  // GET /api/admin/organizations/:orgId/engagement-signals - Get engagement signals for an org
-  apiRouter.get(
-    "/organizations/:orgId/engagement-signals",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const engagementSignals = await orgDb.getEngagementSignals(orgId);
-        res.json(engagementSignals);
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching engagement signals");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch engagement signals",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // SLACK DOMAIN DISCOVERY FOR PROSPECT IDENTIFICATION
-  // =========================================================================
-
-  // GET /api/admin/slack/domains - Get email domains from unmapped Slack users
-  // These are potential organizations to add as prospects
-  apiRouter.get(
-    "/slack/domains",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { min_users, limit, include_free } = req.query;
-
-        const domains = await slackDb.getUnmappedDomains({
-          excludeFreeEmailProviders: include_free !== "true",
-          minUsers: min_users ? parseInt(min_users as string, 10) : 1,
-          limit: limit ? parseInt(limit as string, 10) : 100,
-        });
-
-        // Check which domains already have organizations
-        // First do a bulk lookup from our local organization_domains table (fast)
-        const pool = getPool();
-        const domainList = domains.map((d) => d.domain);
-
-        // Query all matching domains at once
-        const localDomainsResult = await pool.query(
-          `SELECT od.domain, o.workos_organization_id, o.name
-           FROM organization_domains od
-           JOIN organizations o ON o.workos_organization_id = od.workos_organization_id
-           WHERE od.domain = ANY($1)`,
-          [domainList]
-        );
-
-        // Build a map of domain -> org for fast lookup
-        const domainToOrgMap = new Map<string, { id: string; name: string }>();
-        for (const row of localDomainsResult.rows) {
-          domainToOrgMap.set(row.domain, {
-            id: row.workos_organization_id,
-            name: row.name,
-          });
-        }
-
-        // For domains not found locally, check WorkOS (slower, but catches domains
-        // that exist in WorkOS but haven't been synced to our table yet)
-        const missingDomains = domainList.filter((d) => !domainToOrgMap.has(d));
-
-        if (missingDomains.length > 0 && workos) {
-          // Check WorkOS for each missing domain (in parallel, limited batch)
-          await Promise.all(
-            missingDomains.slice(0, 20).map(async (domainName) => {
-              try {
-                const orgs = await workos.organizations.listOrganizations({
-                  limit: 1,
-                  domains: [domainName],
-                });
-                if (orgs.data.length > 0) {
-                  const orgResult = await pool.query(
-                    `SELECT name FROM organizations WHERE workos_organization_id = $1`,
-                    [orgs.data[0].id]
-                  );
-                  if (orgResult.rows.length > 0) {
-                    domainToOrgMap.set(domainName, {
-                      id: orgs.data[0].id,
-                      name: orgResult.rows[0].name,
-                    });
-                  }
-                }
-              } catch {
-                // Ignore WorkOS lookup errors
-              }
-            })
-          );
-        }
-
-        // Enrich domains with org info
-        const enrichedDomains = domains.map((domain) => {
-          const existingOrg = domainToOrgMap.get(domain.domain) || null;
-          return {
-            ...domain,
-            existing_org: existingOrg,
-            is_new_prospect: !existingOrg,
-          };
-        });
-
-        res.json({
-          domains: enrichedDomains,
-          total: enrichedDomains.length,
-          new_prospect_count: enrichedDomains.filter((d) => d.is_new_prospect)
-            .length,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching Slack domains");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch Slack domains",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/slack/domains/:domain/create-prospect - Create a prospect from a Slack domain
-  apiRouter.post(
-    "/slack/domains/:domain/create-prospect",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { domain } = req.params;
-        const { name, prospect_notes } = req.body;
-
-        if (!workos) {
-          return res.status(500).json({ error: "WorkOS not configured" });
-        }
-
-        // Get the users from this domain for context
-        const domainData = await slackDb.getUnmappedDomains({
-          excludeFreeEmailProviders: false,
-          minUsers: 1,
-        });
-        const domainInfo = domainData.find(
-          (d) => d.domain.toLowerCase() === domain.toLowerCase()
-        );
-
-        if (!domainInfo) {
-          return res.status(404).json({
-            error: "Domain not found in unmapped Slack users",
-          });
-        }
-
-        // Generate a name if not provided
-        const orgName =
-          name ||
-          domain
-            .split(".")
-            .slice(0, -1)
-            .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
-            .join(" ");
-
-        // Create organization in WorkOS with the domain
-        const workosOrg = await workos.organizations.createOrganization({
-          name: orgName,
-          domainData: [{ domain: domain, state: DomainDataState.Verified }],
-        });
-
-        logger.info(
-          { orgId: workosOrg.id, name: orgName, domain },
-          "Created WorkOS organization from Slack domain"
-        );
-
-        // Create local record
-        const pool = getPool();
-        const slackUserNames = domainInfo.users
-          .map((u) => u.slack_real_name || u.slack_display_name)
-          .filter(Boolean)
-          .slice(0, 5)
-          .join(", ");
-
-        const notes =
-          prospect_notes ||
-          `Discovered via Slack. ${domainInfo.user_count} user(s) in Slack workspace: ${slackUserNames}`;
-
-        const result = await pool.query(
-          `INSERT INTO organizations (
-            workos_organization_id,
-            name,
-            prospect_status,
-            prospect_source,
-            prospect_notes
-          ) VALUES ($1, $2, $3, $4, $5)
-          RETURNING *`,
-          [workosOrg.id, orgName, "prospect", "slack_discovery", notes]
-        );
-
-        // Auto-enrich the new organization in the background
-        // Don't await - let it run asynchronously
-        enrichOrganization(workosOrg.id, domain).catch((err) => {
-          logger.warn({ err, domain, orgId: workosOrg.id }, "Background enrichment failed");
-        });
-
-        res.status(201).json({
-          ...result.rows[0],
-          domain,
-          slack_users: domainInfo.users,
-          workos_org: {
-            id: workosOrg.id,
-            domains: workosOrg.domains,
-          },
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error creating prospect from Slack domain");
-
-        if (error instanceof Error && error.message.includes("domain")) {
-          return res.status(400).json({
-            error: "Domain error",
-            message: error.message,
-          });
-        }
-
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to create prospect",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // EMAIL CONTACT DOMAIN DISCOVERY
-  // =========================================================================
-
-  // GET /api/admin/email/domains - Get email domains from unmapped email contacts
-  // These are potential organizations to add as prospects (similar to Slack discovery)
-  apiRouter.get(
-    "/email/domains",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { min_users, limit, include_free } = req.query;
-        const excludeFree = include_free !== "true";
-        const minUsers = min_users ? parseInt(min_users as string, 10) : 1;
-        const resultLimit = limit ? parseInt(limit as string, 10) : 100;
-
-        // Common free email providers to exclude
-        const freeEmailDomains = [
-          'gmail.com', 'googlemail.com', 'yahoo.com', 'yahoo.co.uk', 'hotmail.com',
-          'outlook.com', 'live.com', 'msn.com', 'aol.com', 'icloud.com', 'me.com',
-          'mac.com', 'protonmail.com', 'proton.me', 'mail.com', 'zoho.com',
-          'yandex.com', 'gmx.com', 'gmx.net', 'fastmail.com', 'tutanota.com',
-        ];
-
-        const pool = getPool();
-
-        // Build the exclude clause for free email providers
-        let domainExcludeClause = '';
-        const params: (string | number)[] = [];
-        if (excludeFree) {
-          const placeholders = freeEmailDomains.map((_, i) => `$${i + 1}`).join(', ');
-          domainExcludeClause = `AND LOWER(domain) NOT IN (${placeholders})`;
-          params.push(...freeEmailDomains);
-        }
-
-        // Query unmapped email contacts grouped by domain
-        const domainQuery = `
-          SELECT
-            LOWER(domain) as domain,
-            COUNT(*) as user_count,
-            json_agg(json_build_object(
-              'email', email,
-              'display_name', display_name,
-              'email_count', email_count,
-              'last_seen_at', last_seen_at
-            ) ORDER BY email_count DESC) as users
-          FROM email_contacts
-          WHERE mapping_status = 'unmapped'
-            AND domain IS NOT NULL
-            ${domainExcludeClause}
-          GROUP BY LOWER(domain)
-          HAVING COUNT(*) >= $${params.length + 1}
-          ORDER BY COUNT(*) DESC
-          LIMIT $${params.length + 2}
-        `;
-
-        params.push(minUsers, resultLimit);
-        const domainsResult = await pool.query(domainQuery, params);
-
-        // Check which domains already have organizations
-        const enrichedDomains = await Promise.all(
-          domainsResult.rows.map(async (domain) => {
-            let existingOrg = null;
-            try {
-              if (workos) {
-                const orgs = await workos.organizations.listOrganizations({
-                  limit: 1,
-                  domains: [domain.domain],
-                });
-                if (orgs.data.length > 0) {
-                  const orgResult = await pool.query(
-                    `SELECT name, workos_organization_id FROM organizations WHERE workos_organization_id = $1`,
-                    [orgs.data[0].id]
-                  );
-                  if (orgResult.rows.length > 0) {
-                    existingOrg = {
-                      id: orgs.data[0].id,
-                      name: orgResult.rows[0].name,
-                    };
-                  }
-                }
-              }
-            } catch {
-              // Ignore lookup errors
-            }
-
-            return {
-              domain: domain.domain,
-              user_count: parseInt(domain.user_count, 10),
-              users: domain.users,
-              existing_org: existingOrg,
-              is_new_prospect: !existingOrg,
-              source: 'email' as const,
-            };
-          })
-        );
-
-        res.json(enrichedDomains);
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching email domains");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch email domains",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/email/domains/:domain/create-prospect - Create a prospect from an email domain
-  apiRouter.post(
-    "/email/domains/:domain/create-prospect",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { domain } = req.params;
-        const { name, prospect_notes } = req.body;
-
-        if (!workos) {
-          return res.status(500).json({ error: "WorkOS not configured" });
-        }
-
-        const pool = getPool();
-
-        // Get the contacts from this domain for context
-        const contactsResult = await pool.query(
-          `SELECT email, display_name, email_count, last_seen_at
-           FROM email_contacts
-           WHERE mapping_status = 'unmapped'
-             AND LOWER(domain) = LOWER($1)
-           ORDER BY email_count DESC
-           LIMIT 10`,
-          [domain]
-        );
-
-        if (contactsResult.rows.length === 0) {
-          return res.status(404).json({
-            error: "Domain not found in unmapped email contacts",
-          });
-        }
-
-        // Generate a name if not provided
-        const orgName =
-          name ||
-          domain
-            .split(".")
-            .slice(0, -1)
-            .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
-            .join(" ");
-
-        // Create organization in WorkOS with the domain
-        const workosOrg = await workos.organizations.createOrganization({
-          name: orgName,
-          domainData: [{ domain: domain, state: DomainDataState.Verified }],
-        });
-
-        logger.info(
-          { orgId: workosOrg.id, name: orgName, domain },
-          "Created WorkOS organization from email domain"
-        );
-
-        // Create local record
-        const contactNames = contactsResult.rows
-          .map((c) => c.display_name || c.email.split("@")[0])
-          .filter(Boolean)
-          .slice(0, 5)
-          .join(", ");
-
-        const notes =
-          prospect_notes ||
-          `Discovered via email contacts. ${contactsResult.rows.length} contact(s): ${contactNames}`;
-
-        const result = await pool.query(
-          `INSERT INTO organizations (
-            workos_organization_id,
-            name,
-            prospect_status,
-            prospect_source,
-            prospect_notes
-          ) VALUES ($1, $2, $3, $4, $5)
-          RETURNING *`,
-          [workosOrg.id, orgName, "prospect", "email_discovery", notes]
-        );
-
-        // Auto-enrich the new organization in the background
-        enrichOrganization(workosOrg.id, domain).catch((err) => {
-          logger.warn({ err, domain, orgId: workosOrg.id }, "Background enrichment failed");
-        });
-
-        res.status(201).json({
-          ...result.rows[0],
-          domain,
-          email_contacts: contactsResult.rows,
-          workos_org: {
-            id: workosOrg.id,
-            domains: workosOrg.domains,
-          },
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error creating prospect from email domain");
-
-        if (error instanceof Error && error.message.includes("domain")) {
-          return res.status(400).json({
-            error: "Domain error",
-            message: error.message,
-          });
-        }
-
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to create prospect",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // EMAIL CONTACT MANAGEMENT
-  // =========================================================================
-
-  // GET /api/admin/email/contacts - Search and list email contacts
-  apiRouter.get(
-    "/email/contacts",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const pool = getPool();
-        const { search, domain, status, limit: limitParam, offset: offsetParam } = req.query;
-        const resultLimit = limitParam ? parseInt(limitParam as string, 10) : 50;
-        const resultOffset = offsetParam ? parseInt(offsetParam as string, 10) : 0;
-
-        const params: (string | number)[] = [];
-        const conditions: string[] = [];
-
-        if (search && typeof search === 'string') {
-          params.push(`%${search.toLowerCase()}%`);
-          conditions.push(`(LOWER(email) LIKE $${params.length} OR LOWER(display_name) LIKE $${params.length})`);
-        }
-
-        if (domain && typeof domain === 'string') {
-          params.push(domain.toLowerCase());
-          conditions.push(`LOWER(domain) = $${params.length}`);
-        }
-
-        if (status && typeof status === 'string') {
-          params.push(status);
-          conditions.push(`mapping_status = $${params.length}`);
-        }
-
-        const whereClause = conditions.length > 0 ? `WHERE ${conditions.join(' AND ')}` : '';
-
-        // Get total count
-        const countResult = await pool.query(
-          `SELECT COUNT(*) FROM email_contacts ${whereClause}`,
-          params
-        );
-        const total = parseInt(countResult.rows[0].count, 10);
-
-        // Get contacts with their activity count
-        params.push(resultLimit, resultOffset);
-        const contactsResult = await pool.query(
-          `SELECT
-            ec.id,
-            ec.email,
-            ec.display_name,
-            ec.domain,
-            ec.workos_user_id,
-            ec.organization_id,
-            ec.mapping_status,
-            ec.mapping_source,
-            ec.first_seen_at,
-            ec.last_seen_at,
-            ec.email_count,
-            ec.created_at,
-            ec.updated_at,
-            o.name as organization_name,
-            (SELECT COUNT(*) FROM email_activity_contacts eac WHERE eac.contact_id = ec.id) as activity_count
-          FROM email_contacts ec
-          LEFT JOIN organizations o ON ec.organization_id = o.workos_organization_id
-          ${whereClause}
-          ORDER BY ec.last_seen_at DESC NULLS LAST
-          LIMIT $${params.length - 1} OFFSET $${params.length}`,
-          params
-        );
-
-        res.json({
-          contacts: contactsResult.rows,
-          total,
-          limit: resultLimit,
-          offset: resultOffset,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching email contacts");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch email contacts",
-        });
-      }
-    }
-  );
-
-  // GET /api/admin/email/contacts/:id - Get contact details with activity history
-  apiRouter.get(
-    "/email/contacts/:id",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const pool = getPool();
-        const { id } = req.params;
-
-        // Get contact details
-        const contactResult = await pool.query(
-          `SELECT
-            ec.id,
-            ec.email,
-            ec.display_name,
-            ec.domain,
-            ec.workos_user_id,
-            ec.organization_id,
-            ec.mapping_status,
-            ec.mapping_source,
-            ec.first_seen_at,
-            ec.last_seen_at,
-            ec.email_count,
-            ec.mapped_at,
-            ec.mapped_by_user_id,
-            ec.created_at,
-            ec.updated_at,
-            o.name as organization_name,
-            o.company_type as organization_type,
-            o.prospect_status as organization_prospect_status
-          FROM email_contacts ec
-          LEFT JOIN organizations o ON ec.organization_id = o.workos_organization_id
-          WHERE ec.id = $1`,
-          [id]
-        );
-
-        if (contactResult.rows.length === 0) {
-          return res.status(404).json({ error: "Contact not found" });
-        }
-
-        const contact = contactResult.rows[0];
-
-        // Get activity history for this contact
-        const activitiesResult = await pool.query(
-          `SELECT
-            eca.id as activity_id,
-            eca.email_id,
-            eca.message_id,
-            eca.subject,
-            eca.direction,
-            eca.insights,
-            eca.metadata,
-            eca.email_date,
-            eca.created_at,
-            eac.role,
-            eac.is_primary
-          FROM email_contact_activities eca
-          INNER JOIN email_activity_contacts eac ON eac.activity_id = eca.id
-          WHERE eac.contact_id = $1
-          ORDER BY eca.email_date DESC`,
-          [id]
-        );
-
-        res.json({
-          ...contact,
-          activities: activitiesResult.rows,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching email contact details");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch email contact details",
-        });
-      }
-    }
-  );
-
-  // GET /api/admin/email/contacts/by-email/:email - Lookup contact by email address
-  apiRouter.get(
-    "/email/contacts/by-email/:email",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const pool = getPool();
-        const { email } = req.params;
-
-        // Get contact by email
-        const contactResult = await pool.query(
-          `SELECT
-            ec.id,
-            ec.email,
-            ec.display_name,
-            ec.domain,
-            ec.workos_user_id,
-            ec.organization_id,
-            ec.mapping_status,
-            ec.mapping_source,
-            ec.first_seen_at,
-            ec.last_seen_at,
-            ec.email_count,
-            ec.mapped_at,
-            ec.mapped_by_user_id,
-            ec.created_at,
-            ec.updated_at,
-            o.name as organization_name,
-            o.company_type as organization_type,
-            o.prospect_status as organization_prospect_status
-          FROM email_contacts ec
-          LEFT JOIN organizations o ON ec.organization_id = o.workos_organization_id
-          WHERE LOWER(ec.email) = LOWER($1)`,
-          [email]
-        );
-
-        if (contactResult.rows.length === 0) {
-          return res.status(404).json({ error: "Contact not found" });
-        }
-
-        const contact = contactResult.rows[0];
-
-        // Get activity history for this contact
-        const activitiesResult = await pool.query(
-          `SELECT
-            eca.id as activity_id,
-            eca.email_id,
-            eca.message_id,
-            eca.subject,
-            eca.direction,
-            eca.insights,
-            eca.metadata,
-            eca.email_date,
-            eca.created_at,
-            eac.role,
-            eac.is_primary
-          FROM email_contact_activities eca
-          INNER JOIN email_activity_contacts eac ON eac.activity_id = eca.id
-          WHERE eac.contact_id = $1
-          ORDER BY eca.email_date DESC`,
-          [contact.id]
-        );
-
-        res.json({
-          ...contact,
-          activities: activitiesResult.rows,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching email contact by email");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch email contact",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // COMPANY ENRICHMENT (Lusha API)
-  // =========================================================================
-
-  // GET /api/admin/enrichment/status - Check if enrichment is configured
-  apiRouter.get(
-    "/enrichment/status",
-    requireAuth,
-    requireAdmin,
-    async (_req, res) => {
-      res.json({
-        configured: isLushaConfigured(),
-        provider: isLushaConfigured() ? "lusha" : null,
-      });
-    }
-  );
-
-  // POST /api/admin/enrichment/domain/:domain - Enrich a domain with company data
-  apiRouter.post(
-    "/enrichment/domain/:domain",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { domain } = req.params;
-        const { save_to_org_id } = req.body;
-
-        const lusha = getLushaClient();
-        if (!lusha) {
-          return res.status(503).json({
-            error: "Enrichment not configured",
-            message: "LUSHA_API_KEY environment variable not set",
-          });
-        }
-
-        const result = await lusha.enrichCompanyByDomain(domain);
-
-        if (!result.success || !result.data) {
-          return res.status(result.error === "Company not found" ? 404 : 500).json({
-            error: result.error || "Enrichment failed",
-          });
-        }
-
-        const enrichmentData = result.data;
-
-        // Map to our company type
-        const suggestedCompanyType = mapIndustryToCompanyType(
-          enrichmentData.mainIndustry,
-          enrichmentData.subIndustry
-        );
-
-        // Format revenue range if we have raw revenue
-        const revenueRange =
-          enrichmentData.revenueRange || formatRevenueRange(enrichmentData.revenue);
-
-        // If save_to_org_id is provided, save the enrichment data to that org
-        if (save_to_org_id) {
-          const pool = getPool();
-          await pool.query(
-            `UPDATE organizations SET
-              enrichment_data = $1,
-              enrichment_source = 'lusha',
-              enrichment_at = NOW(),
-              enrichment_revenue = $2,
-              enrichment_revenue_range = $3,
-              enrichment_employee_count = $4,
-              enrichment_employee_count_range = $5,
-              enrichment_industry = $6,
-              enrichment_sub_industry = $7,
-              enrichment_founded_year = $8,
-              enrichment_country = $9,
-              enrichment_city = $10,
-              enrichment_linkedin_url = $11,
-              enrichment_description = $12,
-              company_type = COALESCE(company_type, $13),
-              updated_at = NOW()
-            WHERE workos_organization_id = $14`,
-            [
-              JSON.stringify(enrichmentData),
-              enrichmentData.revenue || null,
-              revenueRange,
-              enrichmentData.employeeCount || null,
-              enrichmentData.employeeCountRange || null,
-              enrichmentData.mainIndustry || null,
-              enrichmentData.subIndustry || null,
-              enrichmentData.foundedYear || null,
-              enrichmentData.country || null,
-              enrichmentData.city || null,
-              enrichmentData.linkedinUrl || null,
-              enrichmentData.description || null,
-              suggestedCompanyType,
-              save_to_org_id,
-            ]
-          );
-
-          logger.info(
-            { domain, orgId: save_to_org_id },
-            "Saved enrichment data to organization"
-          );
-        }
-
-        res.json({
-          success: true,
-          domain,
-          enrichment: {
-            companyName: enrichmentData.companyName,
-            description: enrichmentData.description,
-            employeeCount: enrichmentData.employeeCount,
-            employeeCountRange: enrichmentData.employeeCountRange,
-            revenue: enrichmentData.revenue,
-            revenueRange: revenueRange,
-            industry: enrichmentData.mainIndustry,
-            subIndustry: enrichmentData.subIndustry,
-            foundedYear: enrichmentData.foundedYear,
-            country: enrichmentData.country,
-            city: enrichmentData.city,
-            linkedinUrl: enrichmentData.linkedinUrl,
-          },
-          suggested: {
-            companyType: suggestedCompanyType,
-            companyName: enrichmentData.companyName,
-          },
-          creditsUsed: result.creditsUsed,
-          saved: !!save_to_org_id,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error enriching domain");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to enrich domain",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/enrichment/bulk - Enrich multiple domains
-  apiRouter.post(
-    "/enrichment/bulk",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { domains } = req.body;
-
-        if (!Array.isArray(domains) || domains.length === 0) {
-          return res.status(400).json({
-            error: "domains array required",
-          });
-        }
-
-        if (domains.length > 25) {
-          return res.status(400).json({
-            error: "Maximum 25 domains per bulk request",
-          });
-        }
-
-        const lusha = getLushaClient();
-        if (!lusha) {
-          return res.status(503).json({
-            error: "Enrichment not configured",
-            message: "LUSHA_API_KEY environment variable not set",
-          });
-        }
-
-        const results = await lusha.enrichCompaniesInBulk(domains);
-
-        const enrichments: Array<{
-          domain: string;
-          success: boolean;
-          enrichment?: {
-            companyName: string;
-            employeeCount?: number;
-            revenue?: number;
-            industry?: string;
-          };
-          suggestedCompanyType?: string | null;
-          error?: string;
-        }> = [];
-
-        for (const [domain, result] of results) {
-          if (result.success && result.data) {
-            enrichments.push({
-              domain,
-              success: true,
-              enrichment: {
-                companyName: result.data.companyName,
-                employeeCount: result.data.employeeCount,
-                revenue: result.data.revenue,
-                industry: result.data.mainIndustry,
-              },
-              suggestedCompanyType: mapIndustryToCompanyType(
-                result.data.mainIndustry,
-                result.data.subIndustry
-              ),
-            });
-          } else {
-            enrichments.push({
-              domain,
-              success: false,
-              error: result.error,
-            });
-          }
-        }
-
-        res.json({
-          total: domains.length,
-          successful: enrichments.filter((e) => e.success).length,
-          failed: enrichments.filter((e) => !e.success).length,
-          results: enrichments,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error bulk enriching domains");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to bulk enrich domains",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // LUSHA COMPANY PROSPECTING / SEARCH
-  // =========================================================================
-
-  // GET /api/admin/prospecting/filters - Get available filter options
-  apiRouter.get(
-    "/prospecting/filters",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const lusha = getLushaClient();
-        if (!lusha) {
-          return res.status(503).json({
-            error: "Lusha API not configured",
-            message: "Set LUSHA_API_KEY environment variable",
-          });
-        }
-
-        // Fetch all filter options in parallel
-        const [industries, sizes, revenues] = await Promise.all([
-          lusha.getIndustryFilters(),
-          lusha.getCompanySizeFilters(),
-          lusha.getRevenueFilters(),
-        ]);
-
-        res.json({
-          industries,
-          sizes,
-          revenues,
-          // Provide some preset industry groups for ad tech prospecting
-          presets: {
-            adtech: {
-              label: "Ad Tech",
-              description: "Advertising technology, programmatic, DSPs, SSPs",
-              industryKeywords: ["advertising", "marketing", "programmatic"],
-            },
-            agencies: {
-              label: "Agencies",
-              description: "Media agencies, creative agencies, marketing services",
-              industryKeywords: ["agency", "media buying", "creative"],
-            },
-            publishers: {
-              label: "Publishers",
-              description: "Media companies, content publishers, broadcasters",
-              industryKeywords: ["media", "publishing", "broadcasting", "entertainment"],
-            },
-            brands: {
-              label: "Brands",
-              description: "Consumer brands, retail, CPG companies",
-              industryKeywords: ["retail", "consumer", "food", "beverage", "automotive"],
-            },
-          },
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching prospecting filters");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch prospecting filters",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/prospecting/search - Search for companies by criteria
-  apiRouter.post(
-    "/prospecting/search",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const lusha = getLushaClient();
-        if (!lusha) {
-          return res.status(503).json({
-            error: "Lusha API not configured",
-            message: "Set LUSHA_API_KEY environment variable",
-          });
-        }
-
-        const {
-          industryIds,
-          minEmployees,
-          maxEmployees,
-          companySizeIds,
-          revenueIds,
-          countries,
-          states,
-          cities,
-          keywords,
-          page = 1,
-          pageSize = 25,
-        } = req.body;
-
-        const result = await lusha.searchCompanies(
-          {
-            industryIds,
-            minEmployees,
-            maxEmployees,
-            companySizeIds,
-            revenueIds,
-            countries,
-            states,
-            cities,
-            keywords,
-          },
-          page,
-          pageSize
-        );
-
-        if (!result.success) {
-          return res.status(400).json({
-            error: result.error || "Search failed",
-          });
-        }
-
-        // Cross-reference with existing organizations to mark duplicates
-        const pool = getPool();
-        const domains = result.companies
-          .map((c) => c.domain)
-          .filter((d) => d);
-
-        let existingOrgs: Map<string, string> = new Map();
-        if (domains.length > 0) {
-          const existingResult = await pool.query(
-            `SELECT email_domain, name FROM organizations WHERE email_domain = ANY($1)`,
-            [domains]
-          );
-          existingOrgs = new Map(
-            existingResult.rows.map((r) => [r.email_domain, r.name])
-          );
-        }
-
-        // Enrich companies with our classification and existing org status
-        const companies = result.companies.map((company) => {
-          const existingOrgName = company.domain
-            ? existingOrgs.get(company.domain)
-            : null;
-          const suggestedType = mapIndustryToCompanyType(
-            company.mainIndustry,
-            company.subIndustry
-          );
-          const suggestedRevenueRange =
-            company.revenueRange || formatRevenueRange(company.revenue);
-
-          return {
-            ...company,
-            suggestedCompanyType: suggestedType,
-            suggestedRevenueRange,
-            existingOrg: existingOrgName
-              ? { name: existingOrgName, domain: company.domain }
-              : null,
-            isNewProspect: !existingOrgName,
-          };
-        });
-
-        res.json({
-          success: true,
-          companies,
-          total: result.total,
-          page: result.page,
-          pageSize: result.pageSize,
-          hasMore: result.page * result.pageSize < result.total,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error searching companies");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to search companies",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/prospecting/import - Import a company as a prospect
-  apiRouter.post(
-    "/prospecting/import",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { company, autoAssignOwner } = req.body;
-
-        if (!company || !company.domain) {
-          return res.status(400).json({
-            error: "Missing required fields",
-            message: "Company with domain is required",
-          });
-        }
-
-        // Determine company type from industry
-        const companyType = mapIndustryToCompanyType(
-          company.mainIndustry,
-          company.subIndustry
-        );
-
-        // Determine owner (current user if autoAssign)
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        const currentUserId = autoAssignOwner ? (req as any).user?.id : null;
-
-        // Use centralized prospect service to create real WorkOS org
-        const result = await createProspect({
-          name: company.companyName || company.domain,
-          domain: company.domain,
-          company_type: companyType || undefined,
-          prospect_source: "lusha_prospect",
-          prospect_notes: "Imported from Lusha prospecting search",
-          prospect_owner: currentUserId || undefined,
-        });
-
-        if (!result.success) {
-          if (result.alreadyExists) {
-            return res.status(409).json({
-              error: "Organization already exists",
-              existing: result.organization,
-            });
-          }
-          return res.status(400).json({
-            error: "Failed to create prospect",
-            message: result.error,
-          });
-        }
-
-        const orgId = result.organization!.workos_organization_id;
-
-        // Update with enrichment data we already have from Lusha search
-        // (avoids calling Lusha API again since we have the data)
-        const pool = getPool();
-        await pool.query(
-          `UPDATE organizations SET
-            enrichment_data = $1,
-            enrichment_source = 'lusha',
-            enrichment_at = NOW(),
-            enrichment_revenue = $2,
-            enrichment_revenue_range = $3,
-            enrichment_employee_count = $4,
-            enrichment_employee_count_range = $5,
-            enrichment_industry = $6,
-            enrichment_sub_industry = $7,
-            enrichment_founded_year = $8,
-            enrichment_country = $9,
-            enrichment_city = $10,
-            enrichment_linkedin_url = $11,
-            enrichment_description = $12,
-            updated_at = NOW()
-          WHERE workos_organization_id = $13`,
-          [
-            JSON.stringify(company),
-            company.revenue || null,
-            company.revenueRange || formatRevenueRange(company.revenue) || null,
-            company.employeeCount || null,
-            company.employeeCountRange || null,
-            company.mainIndustry || null,
-            company.subIndustry || null,
-            company.foundedYear || null,
-            company.country || null,
-            company.city || null,
-            company.linkedinUrl || null,
-            company.description || null,
-            orgId,
-          ]
-        );
-
-        logger.info(
-          {
-            domain: company.domain,
-            name: company.companyName,
-            orgId,
-          },
-          "Imported prospect from Lusha search"
-        );
-
-        res.json({
-          success: true,
-          organization: result.organization,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error importing prospect");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to import prospect",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // SLACK ACTIVITY INSIGHTS FOR ENGAGEMENT TRACKING
-  // =========================================================================
-
-  // GET /api/admin/slack/activity/top-users - Get most active Slack users
-  apiRouter.get(
-    "/slack/activity/top-users",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { days, limit, mapped_only } = req.query;
-
-        const users = await slackDb.getMostActiveUsers({
-          days: days ? parseInt(days as string, 10) : 30,
-          limit: limit ? parseInt(limit as string, 10) : 50,
-          mappedOnly: mapped_only === "true",
-        });
-
-        res.json({
-          users,
-          total: users.length,
-          period_days: days ? parseInt(days as string, 10) : 30,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching top Slack users");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch top Slack users",
-        });
-      }
-    }
-  );
-
-  // GET /api/admin/slack/activity/user/:slackUserId - Get activity summary for a Slack user
-  apiRouter.get(
-    "/slack/activity/user/:slackUserId",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { slackUserId } = req.params;
-        const { days } = req.query;
-
-        const [summary, mapping] = await Promise.all([
-          slackDb.getActivitySummary(slackUserId, {
-            days: days ? parseInt(days as string, 10) : 30,
-          }),
-          slackDb.getBySlackUserId(slackUserId),
-        ]);
-
-        if (!mapping) {
-          return res.status(404).json({ error: "Slack user not found" });
-        }
-
-        res.json({
-          user: {
-            slack_user_id: mapping.slack_user_id,
-            slack_email: mapping.slack_email,
-            slack_real_name: mapping.slack_real_name,
-            slack_display_name: mapping.slack_display_name,
-            workos_user_id: mapping.workos_user_id,
-            mapping_status: mapping.mapping_status,
-          },
-          activity: summary,
-          period_days: days ? parseInt(days as string, 10) : 30,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching Slack user activity");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch Slack user activity",
-        });
-      }
-    }
-  );
-
-  // GET /api/admin/organizations/:orgId/slack-activity - Get Slack activity for an org
-  apiRouter.get(
-    "/organizations/:orgId/slack-activity",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const { days } = req.query;
-
-        const summary = await slackDb.getOrgActivitySummary(orgId, {
-          days: days ? parseInt(days as string, 10) : 30,
-        });
-
-        res.json({
-          organization_id: orgId,
-          activity: summary,
-          period_days: days ? parseInt(days as string, 10) : 30,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching org Slack activity");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch organization Slack activity",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // ORGANIZATION DOMAINS API
-  // =========================================================================
-
-  // GET /api/admin/organizations/:orgId/domains - List domains for an organization
-  apiRouter.get(
-    "/organizations/:orgId/domains",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const pool = getPool();
-
-        // Get org info
-        const orgResult = await pool.query(
-          `SELECT name, email_domain FROM organizations WHERE workos_organization_id = $1`,
-          [orgId]
-        );
-
-        if (orgResult.rows.length === 0) {
-          return res.status(404).json({ error: "Organization not found" });
-        }
-
-        // Get all domains from our local table
-        const domainsResult = await pool.query(
-          `SELECT domain, is_primary, verified, source, created_at
-           FROM organization_domains
-           WHERE workos_organization_id = $1
-           ORDER BY is_primary DESC, created_at ASC`,
-          [orgId]
-        );
-
-        // Also get domains from WorkOS for comparison
-        let workOSDomains: Array<{ domain: string; state: string }> = [];
-        if (workos) {
-          try {
-            const workOSOrg = await workos.organizations.getOrganization(orgId);
-            workOSDomains = workOSOrg.domains.map((d) => ({
-              domain: d.domain,
-              state: d.state,
-            }));
-          } catch {
-            // Org might not exist in WorkOS
-          }
-        }
-
-        res.json({
-          organization_id: orgId,
-          organization_name: orgResult.rows[0].name,
-          primary_domain: orgResult.rows[0].email_domain,
-          domains: domainsResult.rows,
-          workos_domains: workOSDomains,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error fetching organization domains");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to fetch organization domains",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/organizations/:orgId/domains - Add a domain to an organization
-  apiRouter.post(
-    "/organizations/:orgId/domains",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const { domain, is_primary, sync_to_workos } = req.body;
-
-        if (!domain) {
-          return res.status(400).json({ error: "domain is required" });
-        }
-
-        const normalizedDomain = domain.toLowerCase().trim();
-        const pool = getPool();
-
-        // Verify org exists
-        const orgResult = await pool.query(
-          `SELECT name FROM organizations WHERE workos_organization_id = $1`,
-          [orgId]
-        );
-
-        if (orgResult.rows.length === 0) {
-          return res.status(404).json({ error: "Organization not found" });
-        }
-
-        // Check if domain is already claimed by another org
-        const existingResult = await pool.query(
-          `SELECT od.workos_organization_id, o.name as org_name
-           FROM organization_domains od
-           JOIN organizations o ON o.workos_organization_id = od.workos_organization_id
-           WHERE od.domain = $1`,
-          [normalizedDomain]
-        );
-
-        if (existingResult.rows.length > 0) {
-          const existingOrg = existingResult.rows[0];
-          if (existingOrg.workos_organization_id !== orgId) {
-            return res.status(409).json({
-              error: "Domain already claimed",
-              message: `Domain ${normalizedDomain} is already associated with ${existingOrg.org_name}`,
-              existing_organization_id: existingOrg.workos_organization_id,
-            });
-          }
-          // Domain already belongs to this org
-          return res.json({
-            success: true,
-            message: "Domain already associated with this organization",
-            domain: normalizedDomain,
-          });
-        }
-
-        // If setting as primary, clear existing primary first
-        if (is_primary) {
-          await pool.query(
-            `UPDATE organization_domains SET is_primary = false, updated_at = NOW()
-             WHERE workos_organization_id = $1 AND is_primary = true`,
-            [orgId]
-          );
-        }
-
-        // Insert the domain
-        await pool.query(
-          `INSERT INTO organization_domains (workos_organization_id, domain, is_primary, verified, source)
-           VALUES ($1, $2, $3, false, 'manual')`,
-          [orgId, normalizedDomain, is_primary || false]
-        );
-
-        // If primary, also update the email_domain column
-        if (is_primary) {
-          await pool.query(
-            `UPDATE organizations SET email_domain = $1, updated_at = NOW()
-             WHERE workos_organization_id = $2`,
-            [normalizedDomain, orgId]
-          );
-        }
-
-        // Optionally sync to WorkOS
-        if (sync_to_workos && workos) {
-          try {
-            await workos.organizations.updateOrganization({
-              organization: orgId,
-              domainData: [{ domain: normalizedDomain, state: DomainDataState.Verified }],
-            });
-            // Mark as verified since we added it to WorkOS
-            await pool.query(
-              `UPDATE organization_domains SET verified = true, source = 'workos', updated_at = NOW()
-               WHERE workos_organization_id = $1 AND domain = $2`,
-              [orgId, normalizedDomain]
-            );
-          } catch (workosErr) {
-            logger.warn({ err: workosErr, domain: normalizedDomain }, "Failed to sync domain to WorkOS");
-          }
-        }
-
-        logger.info({ orgId, domain: normalizedDomain, isPrimary: is_primary }, "Added domain to organization");
-
-        res.json({
-          success: true,
-          domain: normalizedDomain,
-          is_primary: is_primary || false,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error adding organization domain");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to add organization domain",
-        });
-      }
-    }
-  );
-
-  // DELETE /api/admin/organizations/:orgId/domains/:domain - Remove a domain from an organization
-  apiRouter.delete(
-    "/organizations/:orgId/domains/:domain",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId, domain } = req.params;
-        const normalizedDomain = domain.toLowerCase().trim();
-        const pool = getPool();
-
-        // Get domain info before deletion
-        const domainResult = await pool.query(
-          `SELECT is_primary, source FROM organization_domains
-           WHERE workos_organization_id = $1 AND domain = $2`,
-          [orgId, normalizedDomain]
-        );
-
-        if (domainResult.rows.length === 0) {
-          return res.status(404).json({ error: "Domain not found for this organization" });
-        }
-
-        const wasPrimary = domainResult.rows[0].is_primary;
-
-        // Delete the domain
-        await pool.query(
-          `DELETE FROM organization_domains WHERE workos_organization_id = $1 AND domain = $2`,
-          [orgId, normalizedDomain]
-        );
-
-        // If we deleted the primary domain, pick a new one
-        if (wasPrimary) {
-          const remaining = await pool.query(
-            `SELECT domain FROM organization_domains
-             WHERE workos_organization_id = $1
-             ORDER BY verified DESC, created_at ASC
-             LIMIT 1`,
-            [orgId]
-          );
-
-          const newPrimary = remaining.rows.length > 0 ? remaining.rows[0].domain : null;
-
-          if (newPrimary) {
-            await pool.query(
-              `UPDATE organization_domains SET is_primary = true, updated_at = NOW()
-               WHERE workos_organization_id = $1 AND domain = $2`,
-              [orgId, newPrimary]
-            );
-          }
-
-          await pool.query(
-            `UPDATE organizations SET email_domain = $1, updated_at = NOW()
-             WHERE workos_organization_id = $2`,
-            [newPrimary, orgId]
-          );
-        }
-
-        logger.info({ orgId, domain: normalizedDomain, wasPrimary }, "Removed domain from organization");
-
-        res.json({
-          success: true,
-          domain: normalizedDomain,
-          was_primary: wasPrimary,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error removing organization domain");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to remove organization domain",
-        });
-      }
-    }
-  );
-
-  // PUT /api/admin/organizations/:orgId/domains/:domain/primary - Set a domain as primary
-  apiRouter.put(
-    "/organizations/:orgId/domains/:domain/primary",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId, domain } = req.params;
-        const normalizedDomain = domain.toLowerCase().trim();
-        const pool = getPool();
-
-        // Verify domain belongs to this org
-        const domainResult = await pool.query(
-          `SELECT domain FROM organization_domains
-           WHERE workos_organization_id = $1 AND domain = $2`,
-          [orgId, normalizedDomain]
-        );
-
-        if (domainResult.rows.length === 0) {
-          return res.status(404).json({ error: "Domain not found for this organization" });
-        }
-
-        // Clear existing primary
-        await pool.query(
-          `UPDATE organization_domains SET is_primary = false, updated_at = NOW()
-           WHERE workos_organization_id = $1 AND is_primary = true`,
-          [orgId]
-        );
-
-        // Set new primary
-        await pool.query(
-          `UPDATE organization_domains SET is_primary = true, updated_at = NOW()
-           WHERE workos_organization_id = $1 AND domain = $2`,
-          [orgId, normalizedDomain]
-        );
-
-        // Update organizations.email_domain
-        await pool.query(
-          `UPDATE organizations SET email_domain = $1, updated_at = NOW()
-           WHERE workos_organization_id = $2`,
-          [normalizedDomain, orgId]
-        );
-
-        logger.info({ orgId, domain: normalizedDomain }, "Set primary domain for organization");
-
-        res.json({
-          success: true,
-          primary_domain: normalizedDomain,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error setting primary domain");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to set primary domain",
-        });
-      }
-    }
-  );
+  // Prospect cleanup routes
+  setupCleanupRoutes(apiRouter);
 
   // =========================================================================
   // USER CONTEXT API (for viewing member context like Addie sees it)
   // =========================================================================
 
   // GET /api/admin/users/:userId/context - Get member context for a user
-  // Accepts either a WorkOS user ID or Slack user ID
   apiRouter.get(
     "/users/:userId/context",
     requireAuth,
@@ -2940,30 +98,23 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
     async (req, res) => {
       try {
         const { userId } = req.params;
-        const { type } = req.query; // 'workos' or 'slack' - defaults to auto-detect
+        const { type } = req.query;
 
         let context;
 
         // Auto-detect or use specified type
         if (type === "slack" || (!type && userId.startsWith("U"))) {
-          // Slack user ID (starts with U)
           context = await getMemberContext(userId);
         } else if (type === "workos" || (!type && userId.startsWith("user_"))) {
-          // WorkOS user ID (starts with user_)
           context = await getWebMemberContext(userId);
         } else {
           // Try both - first check if it's a WorkOS ID
           try {
             context = await getWebMemberContext(userId);
-            // If the context came back with meaningful data, use it
-            if (context.workos_user || context.organization) {
-              // Good, we found the user
-            } else {
-              // Try as Slack ID
+            if (!context.workos_user && !context.organization) {
               context = await getMemberContext(userId);
             }
           } catch {
-            // Fall back to Slack lookup
             context = await getMemberContext(userId);
           }
         }
@@ -3004,7 +155,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           renewals,
           myAccounts,
         ] = await Promise.all([
-          // Needs follow-up: pending next steps due in next 7 days
           pool.query(`
             SELECT COUNT(DISTINCT o.workos_organization_id) as count
             FROM organizations o
@@ -3013,16 +163,12 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
               AND na.next_step_completed_at IS NULL
               AND (na.next_step_due_date IS NULL OR na.next_step_due_date <= NOW() + INTERVAL '7 days')
           `),
-
-          // New signups: created in last 14 days with no activities
           pool.query(`
             SELECT COUNT(*) as count
             FROM organizations o
             WHERE o.created_at > NOW() - INTERVAL '14 days'
               AND NOT EXISTS (SELECT 1 FROM org_activities WHERE organization_id = o.workos_organization_id)
           `),
-
-          // Going cold: last activity > 30 days ago
           pool.query(`
             SELECT COUNT(*) as count
             FROM organizations o
@@ -3034,8 +180,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
                 OR o.subscription_canceled_at IS NOT NULL
               )
           `),
-
-          // Renewals: subscription ending in next 60 days
           pool.query(`
             SELECT COUNT(*) as count
             FROM organizations o
@@ -3044,15 +188,9 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
               AND o.subscription_current_period_end >= NOW()
               AND o.subscription_current_period_end <= NOW() + INTERVAL '60 days'
           `),
-
-          // My accounts: orgs where current user is a stakeholder
           userId
             ? pool.query(
-                `
-            SELECT COUNT(*) as count
-            FROM org_stakeholders
-            WHERE user_id = $1
-          `,
+                `SELECT COUNT(*) as count FROM org_stakeholders WHERE user_id = $1`,
                 [userId]
               )
             : Promise.resolve({ rows: [{ count: 0 }] }),
@@ -3089,7 +227,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
         const { orgId } = req.params;
         const { lookup_key } = req.body;
 
-        // Get the organization details
         const pool = getPool();
         const orgResult = await pool.query(
           `SELECT workos_organization_id, name, is_personal, prospect_contact_email
@@ -3104,31 +241,13 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
         const org = orgResult.rows[0];
         const customerType = org.is_personal ? "individual" : "company";
 
-        // If no specific lookup_key provided, get available products for this customer type
-        let priceId: string | undefined;
-        let selectedProduct: { lookup_key: string; display_name: string; amount_cents: number } | undefined;
+        // Fetch products once - we need the full product object for price_id
+        const products = await getProductsForCustomer({
+          customerType,
+          category: "membership",
+        });
 
-        if (lookup_key) {
-          // Get the specific product
-          const products = await getProductsForCustomer({
-            customerType,
-            category: "membership",
-          });
-          selectedProduct = products.find(p => p.lookup_key === lookup_key);
-          if (!selectedProduct) {
-            return res.status(400).json({
-              error: "Product not found",
-              message: `No product found with lookup key: ${lookup_key}`,
-            });
-          }
-          priceId = selectedProduct.lookup_key;
-        } else {
-          // Return available products for selection
-          const products = await getProductsForCustomer({
-            customerType,
-            category: "membership",
-          });
-
+        if (!lookup_key) {
           return res.json({
             needs_selection: true,
             products: products.map(p => ({
@@ -3141,8 +260,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Look up price ID from lookup key
-        const products = await getProductsForCustomer({ category: "membership" });
         const product = products.find(p => p.lookup_key === lookup_key);
         if (!product) {
           return res.status(400).json({
@@ -3151,7 +268,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Create checkout session
         const baseUrl = process.env.BASE_URL || "https://agenticadvertising.org";
         const session = await createCheckoutSession({
           priceId: product.price_id,
@@ -3221,7 +337,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           billing_address,
         } = req.body;
 
-        // Validate required fields
         if (!lookup_key || !company_name || !contact_name || !contact_email || !billing_address) {
           return res.status(400).json({
             error: "Missing required fields",
@@ -3229,7 +344,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Validate billing address
         if (!billing_address.line1 || !billing_address.city || !billing_address.state ||
             !billing_address.postal_code || !billing_address.country) {
           return res.status(400).json({
@@ -3238,7 +352,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Verify the organization exists
         const pool = getPool();
         const orgResult = await pool.query(
           `SELECT workos_organization_id, name FROM organizations WHERE workos_organization_id = $1`,
@@ -3251,7 +364,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
 
         const org = orgResult.rows[0];
 
-        // Create and send the invoice
         const result = await createAndSendInvoice({
           lookupKey: lookup_key,
           companyName: company_name,
@@ -3275,7 +387,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Update organization to mark invoice was requested
         await pool.query(
           `UPDATE organizations SET
             invoice_requested_at = NOW(),
@@ -3320,253 +431,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
   );
 
   // =========================================================================
-  // AUTOMATIC ENRICHMENT ENDPOINTS
-  // =========================================================================
-
-  // POST /api/admin/enrichment/auto-enrich-domains - Auto-enrich discovered domains
-  // This endpoint enriches domains in batch for the Domain Discovery page
-  apiRouter.post(
-    "/enrichment/auto-enrich-domains",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { domains } = req.body;
-
-        if (!Array.isArray(domains) || domains.length === 0) {
-          return res.json({ results: [] });
-        }
-
-        // Limit to avoid abuse
-        const domainsToEnrich = domains.slice(0, 20);
-
-        const results = await enrichDomainsInBatch(domainsToEnrich);
-
-        // Convert Map to array for JSON response
-        const enrichmentResults = Array.from(results.entries()).map(
-          ([domain, result]) => ({
-            domain,
-            success: result.success,
-            cached: result.cached,
-            data: result.data,
-            error: result.error,
-          })
-        );
-
-        res.json({
-          results: enrichmentResults,
-          total: enrichmentResults.length,
-          successful: enrichmentResults.filter((r) => r.success).length,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error auto-enriching domains");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to auto-enrich domains",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/enrichment/auto-enrich-org/:orgId - Auto-enrich a single organization
-  apiRouter.post(
-    "/enrichment/auto-enrich-org/:orgId",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const { orgId } = req.params;
-
-        const result = await autoEnrichOrganization(orgId);
-
-        if (!result) {
-          return res.status(404).json({
-            error: "No domain found",
-            message: "Organization does not have a domain to enrich",
-          });
-        }
-
-        res.json({
-          success: result.success,
-          cached: result.cached,
-          data: result.data,
-          error: result.error,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error auto-enriching organization");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to auto-enrich organization",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
-  // PROSPECT CLEANUP ENDPOINTS
-  // =========================================================================
-
-  // GET /api/admin/cleanup/status - Check if cleanup is configured
-  apiRouter.get("/cleanup/status", requireAuth, requireAdmin, async (_req, res) => {
-    res.json({
-      configured: isCleanupConfigured(),
-      enrichment_configured: isLushaConfigured(),
-    });
-  });
-
-  // POST /api/admin/cleanup/analyze - Analyze prospects for issues
-  apiRouter.post(
-    "/cleanup/analyze",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const cleanupService = getCleanupService();
-        if (!cleanupService) {
-          return res.status(503).json({
-            error: "Cleanup not configured",
-            message: "ANTHROPIC_API_KEY is required for intelligent cleanup",
-          });
-        }
-
-        const { limit, onlyProblematic, orgIds } = req.body;
-
-        const report = await cleanupService.analyzeProspects({
-          limit: limit || 50,
-          onlyProblematic: onlyProblematic !== false, // Default true
-          orgIds,
-        });
-
-        res.json(report);
-      } catch (error) {
-        logger.error({ err: error }, "Error analyzing prospects");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to analyze prospects",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/cleanup/auto-fix - Auto-fix issues that can be resolved automatically
-  apiRouter.post(
-    "/cleanup/auto-fix",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const cleanupService = getCleanupService();
-        if (!cleanupService) {
-          return res.status(503).json({
-            error: "Cleanup not configured",
-            message: "ANTHROPIC_API_KEY is required for intelligent cleanup",
-          });
-        }
-
-        const { analyses } = req.body;
-
-        if (!Array.isArray(analyses) || analyses.length === 0) {
-          return res.status(400).json({
-            error: "Invalid request",
-            message: "analyses array is required",
-          });
-        }
-
-        const results = await cleanupService.autoFixIssues(analyses);
-
-        res.json({
-          total: results.length,
-          successful: results.filter((r) => r.success).length,
-          results,
-        });
-      } catch (error) {
-        logger.error({ err: error }, "Error auto-fixing prospects");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to auto-fix prospects",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/cleanup/analyze-with-ai/:orgId - Use Claude to analyze a specific prospect
-  apiRouter.post(
-    "/cleanup/analyze-with-ai/:orgId",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const cleanupService = getCleanupService();
-        if (!cleanupService) {
-          return res.status(503).json({
-            error: "Cleanup not configured",
-            message: "ANTHROPIC_API_KEY is required for intelligent cleanup",
-          });
-        }
-
-        const { orgId } = req.params;
-
-        logger.info({ orgId }, "Starting AI analysis for prospect");
-
-        const result = await cleanupService.analyzeWithClaude(orgId);
-
-        res.json(result);
-      } catch (error) {
-        logger.error({ err: error }, "Error in AI analysis");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to analyze prospect with AI",
-        });
-      }
-    }
-  );
-
-  // POST /api/admin/cleanup/batch-analyze-ai - Use Claude to analyze multiple prospects
-  apiRouter.post(
-    "/cleanup/batch-analyze-ai",
-    requireAuth,
-    requireAdmin,
-    async (req, res) => {
-      try {
-        const cleanupService = getCleanupService();
-        if (!cleanupService) {
-          return res.status(503).json({
-            error: "Cleanup not configured",
-            message: "ANTHROPIC_API_KEY is required for intelligent cleanup",
-          });
-        }
-
-        const { orgIds } = req.body;
-
-        if (!Array.isArray(orgIds) || orgIds.length === 0) {
-          return res.status(400).json({
-            error: "Invalid request",
-            message: "orgIds array is required",
-          });
-        }
-
-        // Limit batch size
-        const limitedOrgIds = orgIds.slice(0, 10);
-
-        logger.info(
-          { count: limitedOrgIds.length },
-          "Starting batch AI analysis"
-        );
-
-        const result = await cleanupService.batchAnalyzeWithClaude(limitedOrgIds);
-
-        res.json(result);
-      } catch (error) {
-        logger.error({ err: error }, "Error in batch AI analysis");
-        res.status(500).json({
-          error: "Internal server error",
-          message: "Unable to batch analyze prospects with AI",
-        });
-      }
-    }
-  );
-
-  // =========================================================================
   // WORKOS WIDGET TOKEN API (mounted at /api/admin)
   // =========================================================================
 
@@ -3593,7 +457,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Validate user ID is present
         if (!req.user?.id) {
           return res.status(401).json({
             error: "Authentication required",
@@ -3601,7 +464,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Validate scope
         const validScopes = [
           "widgets:api-keys:manage",
           "widgets:users-table:manage",
@@ -3617,7 +479,6 @@ export function createAdminRouter(): { pageRouter: Router; apiRouter: Router } {
           });
         }
 
-        // Generate widget token
         const token = await workos.widgets.getToken({
           organizationId,
           userId: req.user.id,
