+++ server/src/routes/admin/index.ts
@@ -8,3 +8,10 @@ export { createAdminSlackRouter } from './slack.js';
 export { createAdminEmailRouter } from './email.js';
 export { createAdminFeedsRouter } from './feeds.js';
 export { createAdminNotificationChannelsRouter } from './notification-channels.js';
+
+// Core admin route setup functions
+export { setupProspectRoutes } from './prospects.js';
+export { setupOrganizationRoutes } from './organizations.js';
+export { setupEnrichmentRoutes } from './enrichment.js';
+export { setupDomainRoutes } from './domains.js';
+export { setupCleanupRoutes } from './cleanup.js';
